# perf(cli): forget worker on run