**DO NOT MERGE**

~4% boost on empty script:

```
Benchmark 1: /tmp/deno run /tmp/empty.js
  Time (mean ± σ):      16.6 ms ±   1.0 ms    [User: 11.0 ms, System: 4.1 ms]
  Range (min … max):    14.8 ms …  19.5 ms    178 runs
 
Benchmark 2: target/release/deno run /tmp/empty.js
  Time (mean ± σ):      15.9 ms ±   0.8 ms    [User: 10.7 ms, System: 3.8 ms]
  Range (min … max):    14.5 ms …  18.0 ms    189 runs
 
Summary
  'target/release/deno run /tmp/empty.js' ran
    1.04 ± 0.08 times faster than '/tmp/deno run /tmp/empty.js'
```